## Which problem is this PR solving?
GRPC wants to do all unmarshaling itself based on stock generated code, which doesn't play nicely with our custom, optimized data path.

## Short description of the changes
Replaces the stock generated OTLP trace server with a new one which conforms to GRPC's, shall we say, idiosyncratic interface, but which uses husky's optimized translation instead of unmarshaling. This is a little awkward since (in some cases) we depend on the request info to extract the dataset, so request validation is now happening during Unmarshal, before any grpc interceptors get a look at the request. Fortunately we're not using any interceptors at the moment. The advantage of doing it this early is that we never need to make a copy of the input buffer.

Also moves some testing to work with the full GRPC stack, to remove dependencies on the details of our implementation. Also adds a GRPC concurrency test, which can help expose race conditions.

Benchmark results are promising, note this looks slower than HTTP mainly because of overhead in the benchmark itself, where the GRPC client is re-serializing the request on every iteration:
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkTracesOTLP/grpc          564784        329547        -41.65%
benchmark                         old allocs    new allocs    delta
BenchmarkTracesOTLP/grpc          5070          909           -82.07%
benchmark                         old bytes     new bytes     delta
BenchmarkTracesOTLP/grpc          322256        119653        -62.87%
```